Advices

+++ src/Controller/TaskController.php
@@ -4,6 +4,7 @@ namespace App\Controller;
 
 use App\Entity\Task;
 use App\Form\TaskType;
+use App\Repository\TaskRepository;
 use App\Service\TaskFormHandler;
 use Symfony\Component\Form\Form;
 use Symfony\Component\HttpFoundation\RedirectResponse;
@@ -11,6 +12,7 @@ use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\Security\Csrf\CsrfTokenManagerInterface;
 
 class TaskController extends AbstractController
 {
@@ -19,7 +21,21 @@ class TaskController extends AbstractController
      */
     public function listAction(): Response
     {
+        if (!$this->verifyRole()) {
+            return $this->redirectToRoute('login');
+        }
+
         return $this->render('task/list.html.twig', ['tasks' => $this->getDoctrine()->getRepository('App:Task')->findAll()]);
+    }
+
+    /**
+     * @Route("/tasks/done", name="task_done", methods={"GET"})
+     * @param TaskRepository $taskRepository
+     * @return Response
+     */
+    public function taskIsDoneAction(TaskRepository $taskRepository): Response
+    {
+        return $this->render('task/list.html.twig', ['tasks' => $taskRepository->findTasksIsDone()]);
     }
 
     /**
@@ -31,6 +47,10 @@ class TaskController extends AbstractController
      */
     public function createAction(Request $request, TaskFormHandler $handleForm)
     {
+        if (!$this->verifyRole()) {
+            return $this->redirectToRoute('login');
+        }
+
         $task = new Task($this->getUser());
         /** @var Form $form */
         $form = $this->createForm(TaskType::class, $task);
@@ -54,6 +74,10 @@ class TaskController extends AbstractController
      */
     public function editAction(Request $request, Task $task, TaskFormHandler $handleForm)
     {
+        if (!$this->verifyRole()) {
+            return $this->redirectToRoute('login');
+        }
+
         /** @var Form $form */
         $form = $this->createForm(TaskType::class, $task);
 
@@ -69,15 +93,25 @@ class TaskController extends AbstractController
 
     /**
      * @Route("/tasks/{id}/toggle", name="task_toggle")
+     * @param Request $request
      * @param Task $task
+     * @param CsrfTokenManagerInterface $tokenManager
      * @return RedirectResponse
      */
-    public function toggleTaskAction(Task $task): RedirectResponse
+    public function toggleTaskAction(Request $request, Task $task, CsrfTokenManagerInterface $tokenManager): RedirectResponse
     {
+        if (!$this->verifyRole() || $request->request->get('_token') === null) {
+            return $this->redirectToRoute('login');
+        }
+
+        if (!$request->request->get('_token') || $tokenManager->getToken('toggle'.$task->getId())->getValue() !== $request->request->get('_token')) {
+            return $this->redirectToRoute('logout');
+        }
+
         $task->toggle(!$task->isDone());
         $this->getDoctrine()->getManager()->flush();
 
-        $this->addFlash('success', sprintf('La tâche %s a été marquée comme faite.', $task->getTitle()));
+        $this->addFlash('success', sprintf('Le statut de tâche "%s" a été actualisée.', $task->getTitle()));
 
         return $this->redirectToRoute('task_list');
     }
@@ -85,10 +119,20 @@ class TaskController extends AbstractController
     /**
      * @Route("/tasks/{id}/delete", name="task_delete")
      * @param Task $task
+     * @param Request $request
+     * @param CsrfTokenManagerInterface $tokenManager
      * @return RedirectResponse
      */
-    public function deleteTaskAction(Task $task): RedirectResponse
+    public function deleteTaskAction(Task $task, Request $request, CsrfTokenManagerInterface $tokenManager): RedirectResponse
     {
+        if ($this->unauthorisedDelete($task, $request)) {
+            return $this->redirectToRoute('task_list');
+        }
+
+        if ($tokenManager->getToken('delete'.$task->getId())->getValue() !== $request->request->get('_token')) {
+            return $this->redirectToRoute('logout');
+        }
+
         $em = $this->getDoctrine()->getManager();
         $em->remove($task);
         $em->flush();
@@ -96,5 +140,27 @@ class TaskController extends AbstractController
         $this->addFlash('success', 'La tâche a été supprimée avec succès.');
 
         return $this->redirectToRoute('task_list');
+    }
+
+    private function VerifyRole(): bool
+    {
+        return $this->getUser() ? $this->getUser()->getRole() : false;
+    }
+
+    private function unauthorisedDelete($task, $request): bool
+    {
+        if($request->request->get('_token')) {
+            return false;
+        }
+
+        if (!$task->getUser() && $this->getUser()->getRole()=== 'ROLE_ADMIN') {
+            return false ;
+        }
+
+        if($this->getUser() && $this->getUser()->getId() === $task->getUser()->getId()) {
+            return false;
+        }
+
+        return true ;
     }
 }
